refactor: move to blog module
