+++ src/app/blog/page.tsx
@@ -3,8 +3,9 @@ import { allPosts } from 'contentlayer/generated'
 import type { Metadata } from 'next'
 
 import PageHeader from '@/components/page-header'
-import Posts from '@/components/posts'
 import { getMetadata } from '@/lib/metadata'
+
+import Posts from './posts'
 
 const posts = allPosts
   .sort((a, b) => Number(new Date(b.date)) - Number(new Date(a.date)))
