+++ src/app/blog/posts.tsx
@@ -4,11 +4,10 @@ import type { Post } from 'contentlayer/generated'
 import { m } from 'framer-motion'
 import { useMemo, useState } from 'react'
 
+import { Button, Container, EmptyState, Input } from '@/components/common'
+import { ListBullets, Search, SquaresFour } from '@/components/icons'
+import PostCard from '@/components/post-card'
 import cn from '@/lib/cn'
-
-import { Button, Container, EmptyState, Input } from './common'
-import { ListBullets, Search, SquaresFour } from './icons'
-import PostCard from './post-card'
 
 type LayoutOption = 'list' | 'grid'
